see 11/27 log

+++ bus-gradle-plugin/src/main/java/com/blankj/bus/BusScan.groovy
@@ -1,11 +1,15 @@
 package com.blankj.bus
 
+import com.blankj.util.JavassistUtils
+import com.blankj.util.LogUtils
 import com.blankj.util.ZipUtils
 import com.blankj.utilcode.util.BusUtils
 import groovy.io.FileType
 import javassist.CtClass
 import javassist.CtMethod
 import org.apache.commons.io.FileUtils
+
+import java.lang.reflect.Modifier
 
 class BusScan {
 
@@ -20,9 +24,11 @@ class BusScan {
 
     void scanJar(File jar) {
         File tmp = new File(jar.getParent(), "temp_" + jar.getName())
-        ZipUtils.unzipFile(jar, tmp)
-        scanDir(tmp)
-        FileUtils.forceDelete(tmp)
+        List<File> unzipFile = ZipUtils.unzipFile(jar, tmp)
+        if (unzipFile != null && unzipFile.size() > 0) {
+            scanDir(tmp)
+            FileUtils.forceDelete(tmp)
+        }
     }
 
     void scanDir(File root) {
@@ -47,14 +53,18 @@ class BusScan {
                 // delete .class
                 className = className.substring(0, className.length() - 6)
 
-                CtClass ctClass = Config.mPool.get(className)
+                CtClass ctClass = JavassistUtils.getPool().get(className)
 
                 CtMethod[] methods = ctClass.getDeclaredMethods();
                 for (CtMethod method : methods) {
                     if (method.hasAnnotation(BusUtils.Subscribe)) {
-                        String name = method.getAnnotation(BusUtils.Subscribe).name()
-                        String sign = method.getReturnType().getName() + ' ' + method.getLongName()
-                        busMap.put(name, sign)
+                        if (Modifier.isStatic(method.getModifiers())) {
+                            String name = method.getAnnotation(BusUtils.Subscribe).name()
+                            String sign = method.getReturnType().getName() + ' ' + method.getLongName()
+                            busMap.put(name, sign)
+                        } else {
+                            LogUtils.l(method.getLongName() + "is not static")
+                        }
                     }
                 }
             }
